+++ assets/js/pages/WishItemSearchPage.vue
@@ -1,31 +1,27 @@
 <template>
 
   <div>
-    <!-- Confirm Modal -->
-    <div ref="confirmModalRef" class="modal fade" tabindex="-1">
-      <div class="modal-dialog modal-dialog-centered">
-        <div class="modal-content">
-          <div class="modal-header">
-            <h5 class="modal-title text-primary">Copy confirmation</h5>
-            <button type="button" class="btn-close" @click="confirmInstance.hide()"></button>
-          </div>
-          <div class="modal-body">
-            <p>
-              Are you sure you want to copy this wish to yourself?
-              <strong>{{ itemToCopy?.title }}</strong>?
-            </p>
-          </div>
-          <div class="modal-footer">
-            <button class="btn btn-secondary" @click="confirmInstance.hide()">
-              Cancel
-            </button>
-            <button class="btn btn-accent" @click="confirmCopy">
-              Copy
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
+    <BaseModal
+        ref="confirmModalRef"
+        size="sm"
+        centered="centered"
+    >
+      <template #header>
+        <h5 class="modal-title text-primary">Copy confirmation</h5>
+      </template>
+
+      <p>
+        Are you sure you want to copy this wish to yourself?
+        <strong>{{ itemToCopy?.title }}</strong>?
+      </p>
+
+      <template #footer>
+        <button class="btn btn-accent" @click="confirmCopy">
+          Copy
+        </button>
+      </template>
+
+    </BaseModal>
 
     <!-- Modal backdrop -->
     <div class="modal-backdrop fade show" v-if="showModal"></div>
@@ -50,10 +46,9 @@ import { ref, onMounted } from 'vue'
 import { useWishItemStore } from '@js/stores/WishItemStore'
 import WishItemList from "@js/components/wishitem/WishItemList.vue";
 import { useProfileStore } from "@js/stores/ProfileStore";
-import { Modal } from 'bootstrap'
+import BaseModal from "@js/components/BaseModal.vue";
 
 const confirmModalRef = ref(null) // confirm modal
-let confirmInstance = null
 
 const wishItemStore = useWishItemStore()
 const profileStore = useProfileStore()
@@ -63,7 +58,6 @@ const filters = ref({})
 
 onMounted(async () => {
   wishItemStore.isLoading = true
-  confirmInstance = new Modal(confirmModalRef.value)
 
   await profileStore.fetchMe();
 
@@ -77,7 +71,7 @@ onMounted(async () => {
 
 const handleCopy = async (item) => {
   itemToCopy.value = item;
-  confirmInstance.show();
+  confirmModalRef.value.show();
 }
 
 const confirmCopy = async () => {
@@ -95,7 +89,7 @@ const confirmCopy = async () => {
   }
   delete(clone['@id']);
   await wishItemStore.add(clone);
-  confirmInstance.hide();
+  confirmModalRef.value.hide();
   itemToCopy.value = null;
 
   if(!wishItemStore.error){